Move layers to separate files (#108)
