+++ src/layers/MapLayer.ts
@@ -1,0 +1,7 @@
+import { ReactElement } from 'react'
+
+export interface MapLayer {
+    layer: ReactElement
+    interactiveLayerIds: string[]
+    onClick: (feature: any) => void
+}
